+++ gulpfile.js
@@ -7,7 +7,7 @@ const fs = require('fs');
 // Update the version number in the version.ts file
 gulp.task('versionUpdater', (done) => {
   const pkgContents = require('./package.json');
-  const filePath = './src/common/appInsights/version.ts';
+  const filePath = './src/common/telemetry/version.ts';
   const fileContents = `export const version: string = "{versionPlaceholder}";`;
   const newContents = fileContents.replace("{versionPlaceholder}", pkgContents.version);
   console.log(`Updating version number to: ${pkgContents.version}`);
@@ -21,5 +21,3 @@ const karmaTask = build.karma;
 if (karmaTask) {
   karmaTask.taskConfig.configPath = './config/karma.config.js';
 }
-
-