+++ src/controls/listView/ListView.tsx
@@ -6,7 +6,7 @@ import { findIndex, has, sortBy, isEqual, cloneDeep } from '@microsoft/sp-lodash
 import { FileTypeIcon, IconType } from '../fileTypeIcon/index';
 import * as strings from 'ControlStrings';
 import { IGroupsItems } from './IListView';
-import * as appInsights from '../../common/appInsights';
+import * as telemetry from '../../common/telemetry';
 
 /**
  * File type icon component
@@ -17,7 +17,7 @@ export class ListView extends React.Component<IListViewProps, IListViewState> {
   constructor(props: IListViewProps) {
     super(props);
 
-    appInsights.track('ReactListView', {
+    telemetry.track('ReactListView', {
       viewFields: !!props.viewFields,
       groupByFields: !!props.groupByFields,
       selectionMode: !!props.selectionMode,