+++ src/controls/peoplepicker/PeoplePickerComponent.tsx
@@ -10,7 +10,7 @@ import { ValidationState } from 'office-ui-fabric-react/lib/Pickers';
 import { MessageBar, MessageBarType } from 'office-ui-fabric-react/lib/MessageBar';
 import { SPHttpClient } from '@microsoft/sp-http';
 import styles from './PeoplePickerComponent.module.scss';
-import * as appInsights from '../../common/appInsights';
+import * as telemetry from '../../common/telemetry';
 import {
   assign
 } from 'office-ui-fabric-react/lib/Utilities';
@@ -31,7 +31,7 @@ export class PeoplePicker extends React.Component<IPeoplePickerProps, IPeoplePic
   constructor(props: IPeoplePickerProps) {
     super(props);
 
-    appInsights.track('ReactPeoplePicker', {
+    telemetry.track('ReactPeoplePicker', {
       groupName: !!props.groupName,
       name: !!props.groupName,
       titleText: !!props.titleText
@@ -154,10 +154,16 @@ export class PeoplePicker extends React.Component<IPeoplePickerProps, IPeoplePic
    * On persona item changed event
    */
   private _onPersonItemsChange = (items: any[]) => {
+    const { selectedItems } = this.props;
+
     this.setState({
       selectedPersons: items,
       showmessageerror: items.length > 0 ? false : true
     });
+
+    if (selectedItems) {
+      selectedItems(items);
+    }
   }
 
   /**
@@ -248,8 +254,8 @@ export class PeoplePicker extends React.Component<IPeoplePickerProps, IPeoplePic
   }
 
   /**
-  * Default React component render method
-  */
+   * Default React component render method
+   */
   public render(): React.ReactElement<IPeoplePickerProps> {
     const peoplepicker = (
       <div id="people" className={`${styles.defaultClass} ${this.props.peoplePickerWPclassName ? this.props.peoplePickerWPclassName : ''}`}>
@@ -267,6 +273,7 @@ export class PeoplePicker extends React.Component<IPeoplePickerProps, IPeoplePic
                               'aria-label': 'People Picker'
                             }}
                             itemLimit={this.props.personSelectionLimit || 1}
+                            disabled={this.props.disabled}
                             onChange={this._onPersonItemsChange} />
       </div>
     );